requirements.txt
python-dotenv
pandas
numpy
pymongo[srv]==4.6.1
certifi
scikit-learn
scipy
pyyaml
fastapi
uvicorn
python-multipart
dill

-e .
